refactor logger and add more transports
